rte_virtio: removed virtio_user_dev->max_queues
Generic virtio_dev has the same field already.

The field was previously used to store vhost-user
GET_QUEUE_NUM response, that will be later assigned
to virtio_dev->max_queues via virtio_user_read_dev_config.

This patch removes virtio_user_dev->max_queues
with entire virtio_user_read_dev_config functionality
for virtio_user.

+++ lib/bdev/virtio/rte_virtio/virtio_dev.h
@@ -45,7 +45,10 @@
 struct virtio_dev {
 	struct virtqueue **vqs;
 	uint16_t	started;
-	uint32_t	max_queues;
+
+	/** Max number of queues the host supports. */
+	uint16_t	max_queues;
+
 	uint8_t		port_id;
 	uint64_t	req_guest_features;
 	uint64_t	guest_features;
